+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add all `Operating System` trove classifiers.
+
 ## [4.0.0 (2025-10-21)](https://github.com/kdeldycke/extra-platforms/compare/v3.2.3...v4.0.0)
 
 - Fix detection of GNU/Hurd. Closes #308.